:recycle: Breaks hardcoded codes into components
